Add files via upload
This was an optimization project used for my machine learning certificate from Caltech. In it, I used XGBoost with L1, L2, and no regularization as well as a linear regression model to optimize the validation and cross-validation r^2 score. 

With hyperparameter tuning and regularization, I was able to increase the score by 2.9 pts relative to the default XGBoost model.